libft+ ft_printf tests

+++ libft/ft_pow.c
@@ -6,18 +6,18 @@
 /*   By: cdesvern <cdesvern@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/26 16:33:18 by cdesvern          #+#    #+#             */
-/*   Updated: 2016/03/26 16:37:04 by cdesvern         ###   ########.fr       */
+/*   Updated: 2016/04/05 14:41:28 by cdesvern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_pow(int n, int p)
+uintmax_t	ft_pow(int n, int p)
 {
 	if (p < 0)
-		return ((long)0);
+		return (0);
 	else if (p == 0)
-		return ((long)1);
+		return (1);
 	else
-		return ((long)(n * ft_pow(n, p - 1)));
+		return ((uintmax_t)(n * ft_pow(n, p - 1)));
 }
